refactor: move kubeconfig writing out of hook daemon into standalone script
Move kubeconfig generation from a synchronous call in handler.py to a
standalone script at devinfra/claude/scripts/write_kubeconfig.py,
invoked as a profile background command. This makes the hook daemon
repo-generic — it no longer knows about kubeconfig, K8sConfig, or SOPS
token decryption.

The new script bakes in the cluster defaults (api.allegedly.works,
claude-code-web, claude-sandbox) with optional CLI overrides. It takes
one arg: the output path. Env vars: CLAUDE_PROJECT_DIR + SOPS_AGE_KEY.

Changes:
- New: devinfra/claude/scripts/write_kubeconfig.py (standalone, argparse)
- Delete: hook_daemon/write_kubeconfig_cli.py (replaced)
- Delete: K8sConfig from config.py, k8s: blocks from all profiles
- Delete: write-kubeconfig subcommand from hook_dispatch.py
- Delete: _write_kubeconfig() from handler.py + unused imports
- Update: web + test profiles add kubeconfig bg command (after_env: true)
- Update: claude-sandbox-kubectl-mcp.sh calls the script directly
- Update: container E2E test polls for async kubeconfig, asserts
  baked-in server URL
