+++ devinfra/claude/hook_daemon/BUILD.bazel
@@ -31,15 +31,6 @@ py_library(
     ],
 )
 
-py_test(
-    name = "test_kubeconfig",
-    srcs = ["test_kubeconfig.py"],
-    deps = [
-        ":write_kubeconfig_cli",
-        "@pypi//pytest_bazel",
-    ],
-)
-
 py_library(
     name = "tracing",
     srcs = ["tracing.py"],
@@ -249,27 +240,6 @@ py_binary(
 )
 
 py_library(
-    name = "write_kubeconfig_cli",
-    srcs = ["write_kubeconfig_cli.py"],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":config",
-        "@pypi//pyyaml",
-    ],
-)
-
-py_test(
-    name = "test_write_kubeconfig_cli",
-    srcs = ["test_write_kubeconfig_cli.py"],
-    deps = [
-        ":write_kubeconfig_cli",
-        "@pypi//pytest",
-        "@pypi//pytest_bazel",
-        "@pypi//pyyaml",
-    ],
-)
-
-py_library(
     name = "hook_dispatch_lib",
     srcs = ["hook_dispatch.py"],
     visibility = ["//visibility:public"],
@@ -277,7 +247,6 @@ py_library(
         ":_build_info",
         ":client",
         ":shim_lib",
-        ":write_kubeconfig_cli",
         "//devinfra/claude:session_paths",
         "//devinfra/claude/claude_api/hooks:dispatch_input",
         "@pypi//mako",
