+++ devinfra/claude/hook_daemon/hook_dispatch.py
@@ -20,7 +20,6 @@ from devinfra.claude.hook_daemon._build_info import GIT_COMMIT
 from devinfra.claude.hook_daemon.client import DaemonHttpError, DaemonStartError, call_daemon
 from devinfra.claude.hook_daemon.shim import run_shim
 from devinfra.claude.hook_daemon.shim_install import SHIM_SESSION_ID_ENV
-from devinfra.claude.hook_daemon.write_kubeconfig_cli import main as write_kubeconfig_main
 from devinfra.claude.session_paths import SessionPaths
 
 _adapter: TypeAdapter[AnyHookInput] = TypeAdapter(AnyHookInput)
@@ -70,18 +69,6 @@ def main() -> None:
         # Adjust argv so run_shim sees [shim_name, <original args>]
         sys.argv = [shim_name, *sys.argv[3:]]
         run_shim(shim_name, session_id)
-        return
-
-    # write-kubeconfig subcommand: `claude-hook write-kubeconfig <output-path>`
-    # Used by devinfra/secrets/web_env.sh and claude-sandbox-kubectl-mcp.sh to
-    # materialize a service-account kubeconfig without re-implementing CA /
-    # proxy-url logic in bash. Replaces the now-deleted kube_from_sops.sh.
-    if len(sys.argv) >= 2 and sys.argv[1] == "write-kubeconfig":
-        try:
-            write_kubeconfig_main(sys.argv[2:])
-        except Exception as e:
-            print(f"write-kubeconfig failed: {e}", file=sys.stderr)
-            sys.exit(1)
         return
 
     raw = sys.stdin.buffer.read()
