+++ devinfra/claude/hook_daemon/session_start/container_e2e/test_profile.yaml
@@ -3,9 +3,8 @@
 # Exercises the real web session-start flow end-to-end:
 #   - startup_env_script = real devinfra/secrets/web_env.sh
 #       → decrypts test-encrypted secrets at /project/secrets/*.yaml
-#   - k8s block triggers the daemon's kubeconfig writer, which imports
-#     devinfra.claude.hook_daemon.write_kubeconfig_cli (real code) and
-#     decrypts /project/secrets/claude-web-k8s-token.yaml
+#   - kubeconfig bg command runs the standalone write_kubeconfig.py script,
+#     which decrypts /project/secrets/claude-web-k8s-token.yaml via sops
 #
 # Stripped (vs real web profile):
 #   - setup_tmpfs / setup_docker  (needs mount/privileges)
@@ -18,28 +17,13 @@
 #                                 `--bes_upload_mode=best_effort` flag
 #                                 that used to allow ignoring them was
 #                                 removed — see bazelbuild/bazel#25756).
-#                                 Any alternative would either bake a real
-#                                 BuildBuddy secret into the test or stand
-#                                 up a mock BES receiver; both felt worse
-#                                 than not exercising BES here.
-#   - pre_commit                  (PostToolUse not exercised in v0)
+#   - pre_commit                  (PostToolUse not exercised)
 #   - network-heavy bg commands   (apt-get, pre-commit install, fork remote)
 #   - context_template            (keep output minimal for test assertions)
-#
-# Kept: bazelisk warmup — exercises the full bg-command flow (after_env=true
-# waits for the env file, the bazelisk shim is invoked from the bg command,
-# and the daemon serializes with the subsequent `bazelisk build //:hello`
-# via Bazel's output base lock).
 #
 # The real web profile is at devinfra/claude/hook_daemon/profiles/web/profile.yaml.
 
 startup_env_script: devinfra/secrets/web_env.sh
-
-k8s:
-  server: https://test.example/
-  service_account: claude-code-web
-  service_account_namespace: default
-  namespace: claude-sandbox
 
 idle_watchdog: false
 
@@ -51,6 +35,10 @@ env_exports: |
   export E2E_TEST_MARKER=1
 
 background_commands:
+  - name: kubeconfig
+    command: python3 "$CLAUDE_PROJECT_DIR/devinfra/claude/scripts/write_kubeconfig.py" ~/.kube/config
+    timeout: 30
+    after_env: true
   - name: bazelisk warmup (query //...)
     command: |
       bazelisk query //... >/dev/null &
