+++ devinfra/claude/scripts/BUILD.bazel
@@ -1,0 +1,30 @@
+load("//devinfra/python:defs.bzl", "py_binary", "py_library", "py_test")
+
+exports_files(
+    ["write_kubeconfig.py"],
+    visibility = ["//devinfra/claude:__subpackages__"],
+)
+
+py_library(
+    name = "write_kubeconfig",
+    srcs = ["write_kubeconfig.py"],
+    visibility = ["//visibility:public"],
+    deps = ["@pypi//pyyaml"],
+)
+
+py_binary(
+    name = "write_kubeconfig_bin",
+    main_module = "devinfra.claude.scripts.write_kubeconfig",
+    deps = [":write_kubeconfig"],
+)
+
+py_test(
+    name = "test_write_kubeconfig",
+    srcs = ["test_write_kubeconfig.py"],
+    deps = [
+        ":write_kubeconfig",
+        "@pypi//pytest",
+        "@pypi//pytest_bazel",
+        "@pypi//pyyaml",
+    ],
+)
