+++ devinfra/claude/scripts/test_write_kubeconfig.py
@@ -1,59 +1,75 @@
-"""Smoke test for `claude-hook write-kubeconfig` subcommand."""
+"""Tests for the standalone kubeconfig writer."""
 
 import base64
 import subprocess
 from pathlib import Path
-from textwrap import dedent
 
 import pytest
 import pytest_bazel
 import yaml
 
-from devinfra.claude.hook_daemon import write_kubeconfig_cli
+from devinfra.claude.scripts import write_kubeconfig
 
 
-def _write_profile(path: Path) -> None:
-    path.write_text(
-        dedent("""
-            idle_watchdog: false
-            k8s:
-              server: https://api.example.test:443
-              service_account: claude-code-web
-              service_account_namespace: default
-              namespace: claude-sandbox
-        """).lstrip()
+def test_build_kubeconfig_proxy_url() -> None:
+    kc = write_kubeconfig.build_kubeconfig(
+        token="tok",
+        server="https://k8s.example.com",
+        service_account="test-sa",
+        namespace="secrets-ns",
+        ca_path=None,
+        proxy_url="http://localhost:18081",
     )
+    assert kc["clusters"][0]["cluster"]["proxy-url"] == "http://localhost:18081"
 
 
-def test_write_kubeconfig_end_to_end(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    """`claude-hook write-kubeconfig` writes a kubeconfig with CA + proxy-url set."""
+def test_build_kubeconfig_no_proxy_url() -> None:
+    kc = write_kubeconfig.build_kubeconfig(
+        token="tok",
+        server="https://k8s.example.com",
+        service_account="test-sa",
+        namespace="secrets-ns",
+        ca_path=None,
+        proxy_url=None,
+    )
+    assert "proxy-url" not in kc["clusters"][0]["cluster"]
+
+
+def test_build_kubeconfig_ca_data(tmp_path: Path) -> None:
+    ca_file = tmp_path / "ca.pem"
+    ca_file.write_text("-----BEGIN CERTIFICATE-----\nfake\n-----END CERTIFICATE-----\n")
+    kc = write_kubeconfig.build_kubeconfig(
+        token="tok",
+        server="https://k8s.example.com",
+        service_account="test-sa",
+        namespace="secrets-ns",
+        ca_path=ca_file,
+        proxy_url=None,
+    )
+    assert "certificate-authority-data" in kc["clusters"][0]["cluster"]
+
+
+def test_main_end_to_end(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     project_dir = tmp_path / "repo"
     (project_dir / "secrets").mkdir(parents=True)
     (project_dir / "secrets" / "claude-web-k8s-token.yaml").write_text("stub")
 
-    profile_rel = "devinfra/claude/hook_daemon/profiles/test/profile.yaml"
-    profile_path = project_dir / profile_rel
-    profile_path.parent.mkdir(parents=True)
-    _write_profile(profile_path)
-
     fake_system_ca = tmp_path / "system-ca.pem"
     ca_pem = "-----BEGIN CERTIFICATE-----\nFAKE-CA-DATA\n-----END CERTIFICATE-----\n"
     fake_system_ca.write_text(ca_pem)
-    monkeypatch.setattr(write_kubeconfig_cli, "_SYSTEM_CA_BUNDLE", fake_system_ca)
+    monkeypatch.setattr(write_kubeconfig, "_SYSTEM_CA_BUNDLE", fake_system_ca)
 
     monkeypatch.setenv("CLAUDE_PROJECT_DIR", str(project_dir))
-    monkeypatch.setenv("DUCKTAPE_CLAUDE_HOOKS_PROFILE", profile_rel)
     monkeypatch.setenv("HTTPS_PROXY", "http://egress.example.test:3128")
 
-    # Stub sops so the test doesn't need SOPS_AGE_KEY or a real encrypted file.
     def _fake_sops(cmd: list[str], **kwargs: object) -> subprocess.CompletedProcess[bytes]:
         assert cmd[0] == "sops"
         return subprocess.CompletedProcess(args=cmd, returncode=0, stdout=b"fake-token\n", stderr=b"")
 
-    monkeypatch.setattr(write_kubeconfig_cli.subprocess, "run", _fake_sops)
+    monkeypatch.setattr(write_kubeconfig.subprocess, "run", _fake_sops)
 
     output_path = tmp_path / "out" / "kubeconfig"
-    write_kubeconfig_cli.main([str(output_path)])
+    write_kubeconfig.main([str(output_path), "--server", "https://api.example.test:443"])
 
     kubeconfig = yaml.safe_load(output_path.read_text())
     cluster = kubeconfig["clusters"][0]["cluster"]
@@ -61,26 +77,18 @@ def test_write_kubeconfig_end_to_end(tmp_path: Path, monkeypatch: pytest.MonkeyP
     assert cluster["proxy-url"] == "http://egress.example.test:3128"
     assert cluster["certificate-authority-data"] == base64.b64encode(ca_pem.encode()).decode()
     assert kubeconfig["users"][0]["user"]["token"] == "fake-token"
+    assert kubeconfig["current-context"] == "claude-code-web"
     assert output_path.stat().st_mode & 0o777 == 0o600
 
 
-def test_write_kubeconfig_no_proxy(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    """When HTTPS_PROXY is unset, kubeconfig has no proxy-url (normal on new containers)."""
+def test_main_defaults(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
+    """Uses baked-in defaults when no --server/--service-account/--namespace given."""
     project_dir = tmp_path / "repo"
     (project_dir / "secrets").mkdir(parents=True)
     (project_dir / "secrets" / "claude-web-k8s-token.yaml").write_text("stub")
 
-    profile_rel = "profiles/test/profile.yaml"
-    profile_path = project_dir / profile_rel
-    profile_path.parent.mkdir(parents=True)
-    _write_profile(profile_path)
-
-    fake_system_ca = tmp_path / "system-ca.pem"
-    fake_system_ca.write_text("-----BEGIN CERTIFICATE-----\nSYSTEM\n-----END CERTIFICATE-----\n")
-    monkeypatch.setattr(write_kubeconfig_cli, "_SYSTEM_CA_BUNDLE", fake_system_ca)
-
+    monkeypatch.setattr(write_kubeconfig, "_SYSTEM_CA_BUNDLE", tmp_path / "nonexistent")
     monkeypatch.setenv("CLAUDE_PROJECT_DIR", str(project_dir))
-    monkeypatch.setenv("DUCKTAPE_CLAUDE_HOOKS_PROFILE", profile_rel)
     monkeypatch.delenv("HTTPS_PROXY", raising=False)
     monkeypatch.delenv("https_proxy", raising=False)
     monkeypatch.delenv("HTTP_PROXY", raising=False)
@@ -89,14 +97,15 @@ def test_write_kubeconfig_no_proxy(tmp_path: Path, monkeypatch: pytest.MonkeyPat
     def _fake_sops(cmd: list[str], **kwargs: object) -> subprocess.CompletedProcess[bytes]:
         return subprocess.CompletedProcess(args=cmd, returncode=0, stdout=b"tok", stderr=b"")
 
-    monkeypatch.setattr(write_kubeconfig_cli.subprocess, "run", _fake_sops)
+    monkeypatch.setattr(write_kubeconfig.subprocess, "run", _fake_sops)
 
     output_path = tmp_path / "kubeconfig"
-    write_kubeconfig_cli.main([str(output_path)])
+    write_kubeconfig.main([str(output_path)])
 
     kubeconfig = yaml.safe_load(output_path.read_text())
     cluster = kubeconfig["clusters"][0]["cluster"]
-    assert "certificate-authority-data" in cluster
+    assert cluster["server"] == "https://api.allegedly.works"
+    assert kubeconfig["current-context"] == "claude-code-web"
     assert "proxy-url" not in cluster
